fix mesos healthChecks logic
